+++ public/index.php
@@ -3,5 +3,5 @@
 require_once '../config/dev.php';
 require_once '../vendor/autoload.php';
 session_start();
-$router = new \App\config\Router();
+$router = new \App\src\Router();
 $router->run();
